added sub folders to tests folder rel: #90
