+++ __tests__/global/Link.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Link from '../components/global/Link';
+import Link from '../../components/global/Link';
 import '@testing-library/jest-dom';
 
 const to = '/';
